Updated the following SQL files to do the following
For the tagMatchingInsert.sql, Internal variables and equipment variables are ignored as they do not exist in the Matrikon File

SourceTest.sql

I added an extra condition where I check to see if the DNP3 network name and the DNP3 device name are a substring of the matrikon itemPath

DigitalTest.sql

Added a condition where if the old config tag was a ALA and been updated to a ACM, it will pass this iff  the authorization level is 0.

Added a condition where if the old config tag was a CTV and been updated to a CMD, it will pass if the bit Reserved is "HI"

AnalogMinRatioTest

Added condition to pass if the units are equal to "A" or "kV".