In ComparisonScene Controller, finally udated the code so the user can open up other configs in the last directory where a file was opened.
In comparisonTest, I added an extra insertion for a cleanup job. What it will do is insert tags that are NOT in the new tag config but do exist in the old tag config or the matrikon config (I also added the sql file that I used to carry this out).